qna 3단계 기능구현 - 수정페이지 접근 권한제외

+++ src/main/java/com/ccstudy/qna/service/question/QuestionService.java
@@ -1,27 +1,41 @@
 package com.ccstudy.qna.service.question;
 
+import com.ccstudy.qna.domain.account.AccountRepository;
 import com.ccstudy.qna.domain.question.Question;
 import com.ccstudy.qna.domain.question.QuestionRepository;
+import com.ccstudy.qna.dto.account.AccountSessionDto;
 import com.ccstudy.qna.dto.question.QuestionDetailResponseDto;
 import com.ccstudy.qna.dto.question.QuestionListResponseDto;
 import com.ccstudy.qna.dto.question.QuestionSaveRequestDto;
 import com.ccstudy.qna.dto.question.QuestionUpdateRequestDto;
+import com.ccstudy.qna.error.UnAuthenticationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class QuestionService {
+
     private final QuestionRepository questionRepository;
+
+    private final AccountRepository accountRepository;
 
     //등록하기
     @Transactional
-    public Long save(QuestionSaveRequestDto dto) {
-        return questionRepository.save(dto.toEntity()).getId();
+    public Long save(QuestionSaveRequestDto dto, HttpSession httpSession) {
+        AccountSessionDto accountSessionDto = AccountSessionDto.getInstance(httpSession);
+        Question question = Question.createBuilder()
+                .author(accountRepository.findAccountByEmail(accountSessionDto.getEmail()).orElseThrow(NoSuchElementException::new))
+                .contents(dto.getContents())
+                .title(dto.getTitle())
+                .build();
+        return questionRepository.save(question).getId();
     }
 
     //조회하기
@@ -43,19 +57,29 @@ public class QuestionService {
 
     //게시글 수정
     @Transactional
-    public Long updateQuestion(QuestionUpdateRequestDto dto){
+    public Long updateQuestion(QuestionUpdateRequestDto dto, HttpSession httpSession) {
+        AccountSessionDto accountSessionDto = AccountSessionDto.getInstance(httpSession);
         Question findQuestion = questionRepository.findById(dto.getId())
                 .orElseThrow(IllegalAccessError::new);
-        findQuestion.setTitle(dto.getTitle());
-        findQuestion.setContents(dto.getContents());
-        return findQuestion.getId();
+        if (findQuestion.isCorrectEmail(accountSessionDto.getEmail())) {
+            findQuestion.setTitle(dto.getTitle());
+            findQuestion.setContents(dto.getContents());
+            return findQuestion.getId();
+        }
+        throw new UnAuthenticationException();
     }
 
     //게시글 삭제
     @Transactional
-    public Long delete(Long id){
-        questionRepository.deleteById(id);
-        return id;
+    public Long delete(Long id, HttpSession httpSession) {
+        AccountSessionDto accountSessionDto = AccountSessionDto.getInstance(httpSession);
+        Question findQuestion = questionRepository.findById(id)
+                .orElseThrow(NoSuchElementException::new);
+        if (findQuestion.isCorrectEmail(accountSessionDto.getEmail())) {
+            questionRepository.deleteById(id);
+            return id;
+        }
+        throw new UnAuthenticationException();
     }
 
 }
